Refactor: create RepositoryMethods module
