+++ test/item_repository_test.rb
@@ -58,15 +58,15 @@ RSpec.describe ItemRepository do
   end
 
   it 'returns max item id' do
-    expect(item_repository.max_item_id).to eq(263567474)
+    expect(item_repository.max_id).to eq(263567474)
   end
 
   it 'can create new item instances' do
-    expect(item_repository.max_item_id).to eq(263567474)
+    expect(item_repository.max_id).to eq(263567474)
 
     item_repository.create("SNACK CHEST","A big treasure chest filled with snacks",30,9999)
 
-    expect(item_repository.max_item_id).to eq(263567475)
+    expect(item_repository.max_id).to eq(263567475)
 
     expect(item_repository.find_by_name("SNACK CHEST").id).to eq(263567475)
   end
